# Resolve a SipHash regression from 605